refactor(inventory): extract raw fan-out and SWR cache into lib/inventory
